move SkipLink to new structure
