+++ src/layouts/newLoggedIn.unit.ts
@@ -57,7 +57,7 @@ const setup = () => {
 				},
 			},
 			stubs: {
-				SkipLinks: { template: "<div></div>" },
+				SkipLink: { template: "<div></div>" },
 				"application-error-wrapper": { template: "<div></div>" },
 				snackbar: { template: "<div></div>" },
 				"router-view": { template: "<div></div>" },
