+++ src/modules/ui/skip-link/index.ts
@@ -1,0 +1,3 @@
+import SkipLink from "./SkipLink.vue";
+
+export { SkipLink };
