+++ tsconfig.json
@@ -53,6 +53,7 @@
 			"@ui-light-box": ["src/modules/ui/light-box"],
 			"@ui-preview-image": ["src/modules/ui/preview-image"],
 			"@ui-room-details": ["src/modules/ui/room-details"],
+			"@ui-skip-link": ["src/modules/ui/skip-link"],
 			"@ui-speed-dial-menu": ["src/modules/ui/speed-dial-menu"],
 			"@ui-qr-code": ["src/modules/ui/qr-code"],
 			"@util-board": ["src/modules/util/board"],
