+++ vue.config.js
@@ -70,6 +70,7 @@ module.exports = defineConfig({
 				"@ui-light-box": getDir("src/modules/ui/light-box"),
 				"@ui-preview-image": getDir("src/modules/ui/preview-image"),
 				"@ui-room-details": getDir("src/modules/ui/room-details"),
+				"@ui-skip-link": getDir("src/modules/ui/skip-link"),
 				"@ui-speed-dial-menu": getDir("src/modules/ui/speed-dial-menu"),
 				"@ui-qr-code": getDir("src/modules/ui/qr-code"),
 				"@util-board": getDir("src/modules/util/board"),
